Cccreate

+++ src/main/java/com/galvanize/invoicify/controllers/CompanyController.java
@@ -1,5 +1,6 @@
 package com.galvanize.invoicify.controllers;
 
+import org.springframework.security.core.Authentication;
 import com.galvanize.invoicify.models.Company;
 import com.galvanize.invoicify.models.User;
 import com.galvanize.invoicify.repositories.CompanyRepository;
@@ -13,39 +14,52 @@ public class CompanyController {
 
     private final CompanyRepository companyRepository;
 
-
     public CompanyController(CompanyRepository companyRepository){
         this.companyRepository=companyRepository;
     }
 
     @PostMapping("/api/company")
-    public Company createCompany(@RequestBody Company company) {
-       return this.companyRepository.save(company);
-
+    public Company createCompany(Authentication auth,@RequestBody Company company) {
+       if(auth !=null && auth.isAuthenticated()){
+            return this.companyRepository.save(company);
+        }
+       return null;
     }
     @GetMapping("/api/company/{id}")
-    public Optional<Company> getCompanyById(@PathVariable Long id) {
-     return this.companyRepository.findById(id);
+    public Optional<Company> getCompanyById(Authentication auth, @PathVariable Long id) {
+        if(auth !=null && auth.isAuthenticated()) {
+            return this.companyRepository.findById(id);
+        }
+        return null;
     }
 
    @PutMapping("/api/company/{id}")
-   public Company updateCompany(@PathVariable long id, @RequestBody Company company){
-        Company newCompany =this.companyRepository.findById(id).get();
-        newCompany.setName(company.getName());
-        newCompany.setInvoices(company.getInvoices());
-        return this.companyRepository.save(newCompany);
+   public Company updateCompany(Authentication auth, @PathVariable long id, @RequestBody Company company){
+       if(auth !=null && auth.isAuthenticated()) {
+           Company newCompany = this.companyRepository.findById(id).get();
+           newCompany.setName(company.getName());
+           newCompany.setInvoices(company.getInvoices());
+           return this.companyRepository.save(newCompany);
+       }
+       return null;
    }
 
 
 
     @DeleteMapping("/api/company/{id}")
-    public Optional<Company> deleteCompanyById(@PathVariable Long id) {
-      this.companyRepository.deleteById(id);
-      return this.companyRepository.findById(id); // we want to make sure that id is deleted
+    public Optional<Company> deleteCompanyById(Authentication auth, @PathVariable Long id) {
+        if(auth !=null && auth.isAuthenticated()) {
+            this.companyRepository.deleteById(id);
+            return this.companyRepository.findById(id); // we want to make sure that id is deleted
+        }
+        return null;
     }
     @GetMapping("/api/company")
-    public Iterable<Company> getAll() {
-       return this.companyRepository.findAll();
+    public Iterable<Company> getAll(Authentication auth) {
+        if(auth !=null && auth.isAuthenticated()) {
+            return this.companyRepository.findAll();
+        }
+        return null;
     }
